requirements.txt
watchdog
pushbullet.py